Add inline Style base class and fix IE 11 rendering issue
